refactor: Refactor mixup
#57
